Add migration method to BridgeTree
